MEGA UPDATE!!! Add authorization, authentication; Add admin policy; Add logic for creating and working with databases(user, roles, permissions); And more...

+++ ProductService.BLL/Services/ProductService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Infrastructure;
 using Microsoft.Extensions.Logging;
 using ProductService.BLL.Exceptions;
 using ProductService.Contracts.DTOs;
@@ -8,19 +9,27 @@ using ProductService.DAL.Interfaces;
 
 namespace ProductService.BLL.Services
 {
-    public class ProductService(IProductRepository productRepository, IMapper mapper, ILogger logger) : IProductService
+    public class ProductService(
+        IProductRepository productRepository,
+        IMapper mapper, 
+        ILogger<ProductService> logger) : IProductService
     {
         /// <summary>
         /// Создает продукт.
         /// </summary>
         /// <param name="createProductDTO">DTO с данными для создания продукта.</param>
+        /// <param name="userId">ID авторизованного пользователя отправившего запрос.</param>
         /// <param name="ct">Токен отмены операции.</param>
         /// <returns>DTO продукта(+Id), если создан.</returns>
-        public async Task<ProductDTO> CreateAsync(CreateProductDTO createProductDTO, CancellationToken ct)
+        public async Task<ProductDTO> CreateAsync(CreateProductDTO createProductDTO, int userId, CancellationToken ct)
         {
             logger.LogInformation("[ProductService]CreateAsync TRY");
 
             var mappedProduct = mapper.Map<Product>(createProductDTO);
+
+            mappedProduct.OwnerId = userId;
+            mappedProduct.CreatedDate = DateOnly.FromDateTime(DateTime.UtcNow);
+
             var res = await productRepository.CreateAsync(mappedProduct, ct);
 
             logger.LogInformation("[ProductService]CreateAsync CORRECT");
@@ -32,13 +41,17 @@ namespace ProductService.BLL.Services
         /// Удаляет продукт по идентификатору.
         /// </summary>
         /// <param name="id">ID продукта.</param>
+        /// <param name="userId">ID авторизованного пользователя отправившего запрос.</param>
         /// <param name="ct">Токен отмены операции.</param>
         /// <exception cref="NotFoundException"></exception>
-        public async Task DeleteAsync(int id, CancellationToken ct)
+        public async Task DeleteAsync(int id, int userId ,CancellationToken ct)
         {
             logger.LogInformation("[ProductService]DeleteAsync {id} TRY", id);
 
-            await GetByIdAsync(id, ct);
+            var productDto = await GetByIdAsync(id, ct);
+            if (productDto.OwnerId != userId)
+                throw new UnauthorizedAccessException("Not your product");
+
             await productRepository.DeleteAsync(id, ct);
 
             logger.LogInformation("[ProductService]DeleteAsync id={id} CORRECT", id);
@@ -54,10 +67,10 @@ namespace ProductService.BLL.Services
         {
             logger.LogInformation("[ProductService]GetAllAsync TRY");
 
-            var res = mapper.Map<List<ProductDTO>>(await productRepository.GetAllAsync(ct));
-
+            var products = mapper.Map<List<ProductDTO>>(await productRepository.GetAllAsync(ct));
+                        
             logger.LogInformation("[ProductService]GetAllAsync CORRECT");
-            return res;
+            return products;
         }
 
         /// <summary>
@@ -73,7 +86,7 @@ namespace ProductService.BLL.Services
 
             var product = await productRepository.GetByIdAsync(id, ct);
             if (product == null)
-                throw new NotFoundException($"[ProductService]GetByIdAsync id={id}");
+                throw new NotFoundException($"Product with ID = {id} is not exist");
 
             var mappedProduct = mapper.Map<ProductDTO>(product);
 
@@ -112,14 +125,19 @@ namespace ProductService.BLL.Services
         /// Обновляет продукт.
         /// </summary>
         /// <param name="updateProductDTO">DTO с данными для обновления продукта(может быть не полным)</param>
+        /// <param name="userId">ID авторизованного пользователя отправившего запрос.</param>
         /// <param name="ct">Токен отмены операции.</param>
         /// <returns>DTO продукта, если обновлен; иначе выбрасывает <c>NotFoundException</c>.</returns>
         /// <exception cref="NotFoundException"></exception>
-        public async Task<ProductDTO> UpdateAsync(UpdateProductDTO updateProductDTO, CancellationToken ct)
+        public async Task<ProductDTO> UpdateAsync(UpdateProductDTO updateProductDTO, int userId, CancellationToken ct)
         {
             logger.LogInformation("[ProductService]UpdateAsync TRY");
 
             var product = mapper.Map<Product>(await GetByIdAsync(updateProductDTO.Id, ct)); // тут проверяется его наличие и результат сразу маппится
+
+            if (product.OwnerId != userId)
+                throw new UnauthorizedAccessException("Not your product");
+
             mapper.Map(updateProductDTO, product);
 
             var res = mapper.Map<ProductDTO>(await productRepository.UpdateAsync(product, ct));
@@ -150,6 +168,32 @@ namespace ProductService.BLL.Services
             logger.LogInformation("[ProductService]FindByIdAsync CORRECT");
 
             return mappedProduct;
+        }
+
+        public async Task<List<ProductDTO>> GetByOwnerIdAsync(int ownerId, CancellationToken ct)
+        {
+            logger.LogInformation("[ProductService]GetByOwnerIdAsync TRY ownerId = {ownerId}", ownerId);
+
+            var resItems = await productRepository.GetByOwnerIdAsync(ownerId, ct);
+
+            var mappedList = mapper.Map<List<ProductDTO>>(resItems);
+
+            logger.LogInformation("[ProductService]GetByOwnerIdAsync CORRECT");
+
+            return mappedList;
+        }
+
+        public async Task<List<int>> SetDeletedByOwnerIdAsync(int ownerId, bool isDeleted, CancellationToken ct)
+        {
+            var products = await productRepository.GetByOwnerIdAsync(ownerId, ct);
+
+            foreach(var product in products)
+            {
+                product.IsDeleted = isDeleted;
+                await productRepository.UpdateAsync(product, ct);
+            }
+
+            return products.Select(pr => pr.Id).ToList();
         }
     }
 }
